02_Embeddings_and_RAG/pyproject.toml
[project]
name = "02-embeddings-and-rag"
version = "0.1.0"
description = "Embeddings and RAG"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.0",
    "nump>=5.5.5.5",  # Fix the typo: "nump" -> "numpy"
    "openai>=1.59.7",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.1",
    # Add these new dependencies:
    "sentence-transformers>=2.2.2",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "cohere>=4.0.0",
    "tensorflow-hub>=0.14.0",
    "tensorflow>=2.13.0",
]
